chapter_001/src/test/java/ru/job4j/loop/PaintTest.java
package ru.job4j.loop;

import java.util.StringJoiner;
import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class PaintTest {

    @Test
    public void whenHeight4ThenCreatePyramidWithHeight4() {
        Paint paint = new Paint();
        String rst = paint.piramid(4);
        //System.out.println(rst); //for visual check
        assertThat(rst,
                is(
                        new StringJoiner(System.lineSeparator(), "", System.lineSeparator())
                                .add("   ^   ")
                                .add("  ^^^  ")
                                .add(" ^^^^^ ")
                                .add("^^^^^^^")
                                .toString()
                )
        );
    }
}
